Add files via upload
You need to create a comprehensive class for an n-dimensional vector. Write its test cases too. Like, 2 dimensional and 3 dimensional vectors, the 4 dimensional and 5 dimensional and dimensional vectors also exist, which are natural extensions of the 2 and 3 dimensional vectors. For examples, vector is a tuple 𝑣 = (𝑥0, 𝑥1, 𝑥2, … , 𝑥𝑛−1)
having magnitude and dot product of two n-dimensional vectors are as follows:
|𝑣𝑎. 𝑣𝑏| = 𝑥𝑎0𝑥𝑏0 + 𝑥𝑎1𝑥𝑏1 + 𝑥𝑎2𝑥𝑏2 + ⋯ + 𝑥𝑎(𝑛−1)𝑥𝑏(𝑛−1)
You may even imagine there is one dimensional vector, which in fact are just numbers. Their magnitude and dot product are as √𝑥02 and 𝑥𝑎𝑥𝑏.