Add files via upload
NOTE 1: Separate programming of Task1, Task3 and at least two lists from Task 4 are mandatory for evaluation of the LAB, for Task 2 you may add some member in the database table to facilitate tasks3 and 4. The code discussed in the class for SQLite DB is provided and you may use it for reference.
NOTE 2: You must use the dynamic/prepared SQL statements through programs.

TASK 1: Create a program that displays the following text menu and accept the input accordingly. 
a) Add, s) Search, d) Delete, l) List All, e) Edit, and q) Quit
The valid inputs are a, s, d, l, e, and q. if user enter otherwise, the program informs him, invalid input, choose from below mentioned options and display the menu again. In case the user inputs a valid option, the program performs the appropriate functionality for the books data and displaysthe menu after that. The structure of the courses data is accno: str, title: str, subtitle: str, author: str, coauthors: str, pages: int, price: float, and category: issuable or not issuable. Note: accno is the accession number to access the books should be unique to identify the book record and is different for the same book having multiplecopies in the library.Create a table for the book in an SQLite database named library.db, and later manage the books data in that table.

TASK 2: Manage the member data in the member table in the same database. Members must have the following attributes to manage: membership number, member full name, address, contact number, category (A, B, C or M), membership start date, membership expiry date, membership closing date, and fine paid by him due to late return. membership closing date should be null for an active member and is some past date for a member who left the library and now no book can be issued to him, while member expiry data is a future data. The A category membership is for five years and 10 books are allowed, B category membership is for three years and 7 books are allowed, C category membership is for one year and 3 books are allowed, while the M category membership is for children (minors) and is for one year and 5 books are allowed. The fine for late return is 50 PKR per day for all categories of members.Besides the primary data management, practice to update the closing date and extends the expiry dates for members.

TASK 3: Manage the library register data in the circulation table in the same database. Circulation consists of issuance and return of the books by the members. The circulation table has data for serial number, membership number, accession number, issue date and the return date.
TASK 4: In this task, generate some lists like: category wise active members list, count of each title issued to members, count and total price of each title issued to members,
memberships to be expired in next 10 days, books having only one copy in the library, books having less than 5 copies in the library, members list with their total unpaid fine, and lot more.