Add files via upload
Create a number list class named NUMBERS, to store just number ints or floats. The user of the NUMBERS class, the driving (main) logic, can add a number in the list, delete, alter, search a number in the list, print all of them using the iterators. The catch is to internally store all the numbers in two separate lists for odd and even numbers, but for the users of the NUMBERS class it is one list. The operations results in many numbers, order them in odd even odd even sequence until it is possible. The list also provides the index operator to return the number efficiently.